requirements.txt
streamlit==1.38.0
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
sentence-transformers==3.1.1
spacy==3.7.6
numpy==1.26.4
scikit-learn==1.5.2